Trait fvm::call_manager::CallManager

pub trait CallManager: 'static {
    type Machine: Machine;

    // Required methods
    fn new(
        machine: Self::Machine,
        engine: Engine,
        gas_limit: u64,
        origin: ActorID,
        origin_address: Address,
        receiver: Option<ActorID>,
        receiver_address: Address,
        nonce: u64,
        gas_premium: TokenAmount
    ) -> Self;
    fn send<K: Kernel<CallManager = Self>>(
        &mut self,
        from: ActorID,
        to: Address,
        method: MethodNum,
        params: Option<Block>,
        value: &TokenAmount,
        gas_limit: Option<Gas>,
        read_only: bool
    ) -> Result<InvocationResult>;
    fn finish(self) -> (Result<FinishRet>, Self::Machine);
    fn machine(&self) -> &Self::Machine;
    fn machine_mut(&mut self) -> &mut Self::Machine;
    fn engine(&self) -> &Engine;
    fn gas_tracker(&self) -> &GasTracker;
    fn gas_premium(&self) -> &TokenAmount;
    fn origin(&self) -> ActorID;
    fn next_actor_address(&self) -> Address;
    fn create_actor(
        &mut self,
        code_id: Cid,
        actor_id: ActorID,
        delegated_address: Option<Address>
    ) -> Result<()>;
    fn resolve_address(&self, address: &Address) -> Result<Option<ActorID>>;
    fn set_actor(&mut self, id: ActorID, state: ActorState) -> Result<()>;
    fn get_actor(&self, id: ActorID) -> Result<Option<ActorState>>;
    fn delete_actor(&mut self, id: ActorID) -> Result<()>;
    fn transfer(
        &mut self,
        from: ActorID,
        to: ActorID,
        value: &TokenAmount
    ) -> Result<()>;
    fn nonce(&self) -> u64;
    fn invocation_count(&self) -> u64;
    fn limiter_mut(&mut self) -> &mut <Self::Machine as Machine>::Limiter;
    fn append_event(&mut self, evt: StampedEvent);

    // Provided methods
    fn price_list(&self) -> &PriceList { ... }
    fn context(&self) -> &MachineContext { ... }
    fn blockstore(&self) -> &<Self::Machine as Machine>::Blockstore { ... }
    fn externs(&self) -> &<Self::Machine as Machine>::Externs { ... }
    fn charge_gas(&self, charge: GasCharge) -> Result<GasTimer> { ... }

}

The CallManager manages a single call stack.

When a top-level message is executed:

The crate::executor::Executor creates a CallManager for that message, giving itself to the CallManager.

The crate::executor::Executor calls the specified actor/method using CallManager::send().

The CallManager then constructs a Kernel and executes the actual actor code on that kernel.

If an actor calls another actor, the Kernel will:

Detach the CallManager from itself.

Call CallManager::send() to execute the new message.

Re-attach the CallManager.

Return.

Required Associated Types

type Machine: Machine

The underlying Machine on top of which this CallManager executes.

Required Methods

fn new( machine: Self::Machine, engine: Engine, gas_limit: u64, origin: ActorID, origin_address: Address, receiver: Option<ActorID>, receiver_address: Address, nonce: u64, gas_premium: TokenAmount ) -> Self

Construct a new call manager.

fn send<K: Kernel<CallManager = Self>>( &mut self, from: ActorID, to: Address, method: MethodNum, params: Option<Block>, value: &TokenAmount, gas_limit: Option<Gas>, read_only: bool ) -> Result<InvocationResult>

Send a message. The type parameter K specifies the the kernel on top of which the target actor should execute.

fn finish(self) -> (Result<FinishRet>, Self::Machine)

Finishes execution, returning the gas used, machine, and exec trace if requested.

fn machine(&self) -> &Self::Machine

Returns a reference to the machine.

fn machine_mut(&mut self) -> &mut Self::Machine

Returns a mutable reference to the machine.

fn engine(&self) -> &Engine

Returns a reference to the engine

fn gas_tracker(&self) -> &GasTracker

Returns a reference to the gas tracker.

fn gas_premium(&self) -> &TokenAmount

Returns the gas premium paid by the currently executing message.

fn origin(&self) -> ActorID

Getter for origin actor.

fn next_actor_address(&self) -> Address

Get the actor address (f2) that will should be assigned to the next actor created.

This method doesn’t have any side-effects and will continue to return the same address until create_actor is called next.

fn create_actor( &mut self, code_id: Cid, actor_id: ActorID, delegated_address: Option<Address> ) -> Result<()>

Create a new actor with the given code CID, actor ID, and delegated address. This method does not register the actor with the init actor. It just creates it in the state-tree.

It handles all appropriate gas charging for creating new actors.

fn resolve_address(&self, address: &Address) -> Result<Option<ActorID>>

Resolve an address into an actor ID, charging gas as appropriate.

fn set_actor(&mut self, id: ActorID, state: ActorState) -> Result<()>

Sets an actor in the state-tree, charging gas as appropriate. Use create_actor if you want to create a new actor.

fn get_actor(&self, id: ActorID) -> Result<Option<ActorState>>

Looks up an actor in the state-tree, charging gas as appropriate.

fn delete_actor(&mut self, id: ActorID) -> Result<()>

Deletes an actor from the state-tree, charging gas as appropriate.

fn transfer( &mut self, from: ActorID, to: ActorID, value: &TokenAmount ) -> Result<()>

Transfers tokens from one actor to another, charging gas as appropriate.

fn nonce(&self) -> u64

Getter for message nonce.

fn invocation_count(&self) -> u64

Gets the total invocations done on this call stack.

fn limiter_mut(&mut self) -> &mut <Self::Machine as Machine>::Limiter

Limit memory usage throughout a message execution.

fn append_event(&mut self, evt: StampedEvent)

Appends an event to the event accumulator.

Provided Methods

fn price_list(&self) -> &PriceList

Returns the current price list.

fn context(&self) -> &MachineContext

Returns the machine context.

fn blockstore(&self) -> &<Self::Machine as Machine>::Blockstore

Returns the blockstore.

fn externs(&self) -> &<Self::Machine as Machine>::Externs

Returns the externs.

fn charge_gas(&self, charge: GasCharge) -> Result<GasTimer>

Charge gas.