Struct fvm::gas::PriceList

pub struct PriceList { /* private fields */ }

Provides prices for operations in the VM. All costs are in milligas.

Implementations

impl PriceList

pub fn on_chain_message(&self, msg_size: usize) -> GasCharge

Returns the gas required for storing a message of a given size in the chain, plus the cost of updating the sending actor’s nonce and balance in the state-tree.

pub fn on_value_transfer(&self) -> GasCharge

Returns the gas required when invoking a method.

pub fn on_method_invocation( &self, _param_size: u32, param_links: usize ) -> GasCharge

Returns the gas required when invoking a method.

pub fn on_method_return( &self, call_depth: u32, return_size: u32, return_links: usize ) -> GasCharge

Returns the gas required for returning a value from a method. At the top-level, this charges for storing the block on-chain. Everywhere else, it charges for tracking IPLD links.

pub fn on_syscall(&self) -> GasCharge

Returns the gas cost to be applied on a syscall.

pub fn on_create_actor(&self, new_address: bool) -> GasCharge

Returns the gas required for creating an actor. Pass true to when explicitly assigning a new address.

pub fn on_delete_actor(&self) -> GasCharge

Returns the gas required for deleting an actor.

pub fn on_verify_signature( &self, sig_type: SignatureType, data_len: usize ) -> GasCharge

Returns gas required for signature verification.

pub fn on_recover_secp_public_key(&self) -> GasCharge

Returns gas required for recovering signer pubkey from signature

pub fn on_hashing(&self, hasher: SupportedHashes, data_len: usize) -> GasCharge

Returns gas required for hashing data.

pub fn on_utf8_validation(&self, len: usize) -> GasCharge

pub fn on_compute_unsealed_sector_cid( &self, _proof: RegisteredSealProof, _pieces: &[PieceInfo] ) -> GasCharge

Returns gas required for computing unsealed sector Cid.

pub fn on_verify_seal(&self, _info: &SealVerifyInfo) -> GasCharge

Returns gas required for seal verification.

pub fn on_verify_aggregate_seals( &self, aggregate: &AggregateSealVerifyProofAndInfos ) -> GasCharge

pub fn on_verify_replica_update( &self, _replica: &ReplicaUpdateInfo ) -> GasCharge

Returns gas required for replica verification.

pub fn on_verify_post(&self, info: &WindowPoStVerifyInfo) -> GasCharge

Returns gas required for PoSt verification.

pub fn on_verify_consensus_fault( &self, _h1_len: usize, _h2_len: usize, _extra_len: usize ) -> GasCharge

Returns gas required for verifying consensus fault.

pub fn on_get_randomness(&self, lookback: ChainEpoch) -> GasCharge

Returns the cost of the gas required for getting randomness from the client with the given lookback.

pub fn on_block_open_base(&self) -> GasCharge

Returns the base gas required for loading an object, independent of the object’s size.

pub fn on_block_open(&self, data_size: usize, links: usize) -> GasCharge

Returns the gas required for loading an object based on the size of the object.

pub fn on_block_read(&self, data_size: usize) -> GasCharge

Returns the gas required for reading a loaded object.

pub fn on_block_create(&self, data_size: usize, links: usize) -> GasCharge

Returns the gas required for adding an object to the FVM cache.

pub fn on_block_link( &self, hash_code: SupportedHashes, data_size: usize ) -> GasCharge

Returns the gas required for committing an object to the state blockstore.

pub fn on_block_stat(&self) -> GasCharge

Returns the gas required for storing an object.

pub fn on_actor_lookup(&self) -> GasCharge

Returns the gas required to lookup an actor in the state-tree.

pub fn on_actor_update(&self) -> GasCharge

Returns the gas required to update an actor in the state-tree. Assumes that the actor lookup fee has already been charged.

pub fn on_actor_create(&self) -> GasCharge

Returns the gas required to create a new actor in the state-tree. Assumes that the actor lookup and update fees have already been charged.

pub fn on_self_balance(&self) -> GasCharge

Returns the gas required for accessing the balance of the current actor.

pub fn on_balance_of(&self) -> GasCharge

Returns the gas required for accessing the balance of an actor.

pub fn on_resolve_address(&self) -> GasCharge

Returns the gas required for resolving an actor address.

Might require lookup in the state tree as well as loading the state of the init actor.

pub fn on_lookup_delegated_address(&self) -> GasCharge

Returns the gas required for looking up an actor’s delegated address.

pub fn on_get_actor_code_cid(&self) -> GasCharge

Returns the gas required for getting the CID of the code of an actor.

Might require looking up the actor in the state tree.

pub fn on_get_builtin_actor_type(&self) -> GasCharge

Returns the gas required for looking up the type of a builtin actor by CID.

pub fn on_get_code_cid_for_type(&self) -> GasCharge

Returns the gas required for looking up the CID of a builtin actor by type.

pub fn on_tipset_cid(&self, lookback: ChainEpoch) -> GasCharge

Returns the gas required for looking up a tipset CID with the given lookback.

pub fn on_network_context(&self) -> GasCharge

Returns the gas required for accessing the network context.

pub fn on_message_context(&self) -> GasCharge

Returns the gas required for accessing the message context.

pub fn init_memory_gas(&self, min_memory_bytes: usize) -> Gas

Returns the gas required for initializing memory.

pub fn grow_memory_gas(&self, grow_memory_bytes: usize) -> Gas

Returns the gas required for growing memory.

pub fn init_table_gas(&self, min_table_elements: u32) -> Gas

Returns the gas required for initializing tables.

pub fn on_actor_event( &self, entries: usize, keysize: usize, valuesize: usize ) -> GasCharge

pub fn on_get_root(&self) -> GasCharge

pub fn on_set_root(&self) -> GasCharge