Module fvm::kernel::prelude

Import this module (with a glob) if you’re implementing a kernel, especially if you want to use ambassador to delegate the implementation.

Re-exports

pub use super::ActorOps;

pub use super::CryptoOps;

pub use super::DebugOps;

pub use super::EventOps;

pub use super::IpldBlockOps;

pub use super::MessageOps;

pub use super::NetworkOps;

pub use super::RandomnessOps;

pub use super::SelfOps;

pub use super::SendOps;

pub use super::UpgradeOps;

pub use super::Block;

pub use super::BlockId;

pub use super::BlockRegistry;

pub use super::BlockStat;

pub use super::CallResult;

pub use super::Kernel;

pub use super::SyscallHandler;

pub use crate::gas::Gas;

pub use crate::gas::GasTimer;

pub use crate::gas::PriceList;

Macros

ambassador_impl_ActorOps

A macro to be used by ambassador::Delegate to delegate ActorOps

ambassador_impl_CryptoOps

A macro to be used by ambassador::Delegate to delegate CryptoOps

ambassador_impl_DebugOps

A macro to be used by ambassador::Delegate to delegate DebugOps

ambassador_impl_EventOps

A macro to be used by ambassador::Delegate to delegate EventOps

ambassador_impl_IpldBlockOps

A macro to be used by ambassador::Delegate to delegate IpldBlockOps

ambassador_impl_MessageOps

A macro to be used by ambassador::Delegate to delegate MessageOps

ambassador_impl_NetworkOps

A macro to be used by ambassador::Delegate to delegate NetworkOps

ambassador_impl_RandomnessOps

A macro to be used by ambassador::Delegate to delegate RandomnessOps

ambassador_impl_SelfOps

A macro to be used by ambassador::Delegate to delegate SelfOps

ambassador_impl_SendOps

A macro to be used by ambassador::Delegate to delegate SendOps

ambassador_impl_UpgradeOps

A macro to be used by ambassador::Delegate to delegate UpgradeOps

Structs

Address

Address is the struct that defines the protocol and data payload conversion from either a public key or value

ExitCode

ExitCode defines the exit code from the VM invocation.

MessageContext

NetworkContext

NetworkVersion

Specifies the network version

SendFlags

Flags passed to the send syscall.

TokenAmount

A quantity of native tokens. A token amount is an integer, but has a human interpretation as a value with 18 decimal places. This is a new-type in order to prevent accidental conversion from other BigInts. From/Into BigInt is missing by design.

Enums

SignatureType

Signature variants for Filecoin signatures.

Constants

RANDOMNESS_LENGTH

SECP_PUB_LEN

Secp256k1 Public key length in bytes.

SECP_SIG_LEN

Secp256k1 signature length in bytes.

SECP_SIG_MESSAGE_HASH_SIZE

Length of the signature input message hash in bytes (32).

Type Aliases

ActorID

Identifier for Actors, includes builtin and initialized actors

ChainEpoch

Epoch number of a chain. This acts as a proxy for time within the VM.

Cid

A Cid that contains a multihash with an allocated size of 512 bits.

MethodNum

Method number indicator for calling actor methods.

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.

Derive Macros

Delegate

Delegate the implementation of a trait to a struct field/enum variants by adding #[derive(Delegate)] and its associated attribute #[delegate(Trait)] to it:

Multihash