Module fvm::syscalls

Structs

InvocationData

Invocation data attached to a wasm “store” and available to the syscall binding.

Functions

bind_syscalls

charge_for_exec

Updates the FVM-side gas tracker with newly accrued execution gas charges.

charge_for_init

Charge for the initial memory and tables before a Wasm module is instantiated.

record_init_time

Record the time it took to initialize a module.

update_gas_available

Updates the global available gas in the Wasm module after a syscall, to account for any gas consumption that happened on the host side.