Trait fvm::syscalls::Syscall

pub trait Syscall<K, Args, Ret>: Send + Sync + 'static {
    // Required method
    fn link(
        self,
        linker: &mut Linker<K>,
        module: &'static str,
        name: &'static str
    ) -> Result<()>;
}

A Syscall is a function in the form fn(Context<'_, K>, I...) -> R where:

K is the kernel type. Constrain this to the precise kernel operations you need, or even a specific kernel implementation.

I..., the syscall parameters, are 0-8 types, each one of u32, u64, i32, or i64.

R is a type implementing IntoControlFlow. This is usually one of:

kernel::Result<T> or ControlFlow<T> where T, the return value type, is SyscallSafe.

Abort for syscalls that only abort (revert) the currently running actor.

You generally shouldn’t implement this trait yourself.

Required Methods

fn link( self, linker: &mut Linker<K>, module: &'static str, name: &'static str ) -> Result<()>

Link this syscall with the specified linker, module name, and function name.