Module fvm::gas

Structs

Gas

A typesafe representation of gas (internally stored as milligas).

GasCharge

Single gas charge in the VM. Contains information about what gas was for, as well as the amount of gas needed for computation and storage respectively.

GasDuration

Data structure to encapsulate the optional duration which is set by the GasTimer.

GasTimer

A handle returned by charge_gas which must be used to mark the end of the execution associated with that gas.

GasTracker

PriceList

Provides prices for operations in the VM. All costs are in milligas.

WasmGasPrices

Constants

MILLIGAS_PRECISION

Functions

price_list_by_network_version

Returns gas price list by NetworkVersion for gas consumption.

Type Aliases

GasInstant

Type alias so that we can disable this with a compiler flag.